sync_monitor/requirements.txt
pyzmq>=26.0.0
yfinance>=0.2.0
